first_bot/requirements.txt
pandas==2.1.4
numpy==1.26.2
python-binance==1.0.19
matplotlib==3.8.2
colorama>=0.4.4
python-dotenv>=0.21.0
websocket-client>=1.3.1
requests>=2.27.1
seaborn==0.13.0
tqdm>=4.66.1
psutil>=5.9.8
json5>=0.9.14
python-dateutil>=2.8.2

# Nota: O pacote TA-Lib precisa ser instalado separadamente
# Windows: Baixar o arquivo .whl em https://www.lfd.uci.edu/~gohlke/pythonlibs/#ta-lib
# Linux/Mac: Instruções em https://github.com/mrjbq7/ta-lib#installation 